data/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
gradio==5.8.0
gradio_client==1.5.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.2.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
pycocotools==2.0.8
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.8.2
safehttpx==0.1.6
scikit-learn==1.6.0
scipy==1.14.1
-e git+https://github.com/deep-overflow/mug.git@e99aa4613f96ca3289fb5742dc701b37780aba8d#egg=segment_anything&subdirectory=segment-anything
semantic-version==2.10.0
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tomlkit==0.13.2
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tqdm==4.67.1
triton==3.0.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.1
wheel==0.44.0
